fix(paasns): race condition during capability secret creation
When retrieving secrets for a capability, the PaasNS controller pulls
all secrets from the Paas capabilities specification, as well as from
the PaasNS spec. But the PaasNS for the capability itself defines its
secret based on the capability specification, so this guarantees
duplicate secret definitions. The ensuing create-or-update loop
iterating over the list of secrets then contains a race condition where
the creation of the first secret must be completed by Kubernetes before
the second iteration updates it; otherwise the duplicate will attempt to
be created resulting in failure.

The fix is to deduplicate the secrets list prior to creation or updates.